yess-backend/pyproject.toml
[tool.poetry]
name = "yess-loyalty-backend"
version = "1.0.0"
description = "Backend для системы лояльности YESS"
authors = ["YESS Team <dev@yess-loyalty.kg>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.22.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.10"}
alembic = "^1.10.3"
psycopg2-binary = "^2.9.6"
pydantic = "^2.6.4"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
requests = "^2.28.2"
geopy = "^2.3.0"
python-dotenv = "^1.0.0"
pillow = "^9.5.0"
redis = "^4.5.4"
prometheus-client = "^0.16.0"
locust = "^2.15.1"
sentry-sdk = "^1.21.1"
loguru = "^0.7.0"
httpx = "^0.24.0"
geoalchemy2 = "^0.14.2"
psutil = "^5.9.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
locust = "^2.15.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
lines_after_imports = 2

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
strict_optional = true
